Move slice fetching to poststack handle (2/8)
Moving the fetching of slices into the class handling the poststack
data. The fetching of slice metadata is note moved yet. This leads to
some duplication of helper functions in `vds.cpp` and `poststack.cpp`
which will be resolved when the fetching of metadata is also moved.
